Move fault injection hardening API to new directory
To facilitate splitting it into a stand-alone repo.

Fixes #2798
