+++ boot/boot_serial/src/boot_serial_encryption.c
@@ -14,7 +14,7 @@
 #include <../src/bootutil_priv.h>
 #include "bootutil/bootutil_log.h"
 #include "bootutil/bootutil_public.h"
-#include "bootutil/fault_injection_hardening.h"
+#include "fault_injection_hardening.h"
 
 BOOT_LOG_MODULE_REGISTER(serial_encryption);
