+++ boot/bootutil/include/bootutil/image.h
@@ -30,7 +30,7 @@
 
 #include <inttypes.h>
 #include <stdbool.h>
-#include "bootutil/fault_injection_hardening.h"
+#include "fault_injection_hardening.h"
 
 #ifdef __cplusplus
 extern "C" {
