+++ boot/bootutil/src/fault_injection_hardening_delay_rng_mbedtls.c
@@ -4,7 +4,7 @@
  * Copyright (c) 2020 Arm Limited
  */
 
-#include "bootutil/fault_injection_hardening.h"
+#include "fault_injection_hardening.h"
 
 #ifdef FIH_ENABLE_DELAY
