+++ boot/espressif/main.c
@@ -6,7 +6,7 @@
 
 #include <bootutil/bootutil.h>
 #include <bootutil/bootutil_log.h>
-#include <bootutil/fault_injection_hardening.h>
+#include <fault_injection_hardening.h>
 #include <bootutil/image.h>
 
 #include "bootloader_init.h"
