+++ boot/fault_injection_hardening/include/fault_injection_hardening_delay_rng.h
