+++ boot/zephyr/firmware_loader.c
@@ -13,7 +13,7 @@
 #include "bootutil_priv.h"
 #include "bootutil/bootutil_log.h"
 #include "bootutil/bootutil_public.h"
-#include "bootutil/fault_injection_hardening.h"
+#include "fault_injection_hardening.h"
 
 #include "io/io.h"
 #include "mcuboot_config/mcuboot_config.h"
