+++ boot/zephyr/hooks_sample.c
@@ -18,7 +18,7 @@
 #include <zephyr/kernel.h>
 #include "bootutil/image.h"
 #include "bootutil/bootutil.h"
-#include "bootutil/fault_injection_hardening.h"
+#include "fault_injection_hardening.h"
 #include "flash_map_backend/flash_map_backend.h"
 
 /* @retval 0: header was read/populated
