+++ boot/zephyr/single_loader.c
@@ -12,7 +12,7 @@
 #include "bootutil/bootutil.h"
 #include "bootutil/bootutil_log.h"
 #include "bootutil/bootutil_public.h"
-#include "bootutil/fault_injection_hardening.h"
+#include "fault_injection_hardening.h"
 
 #include "mcuboot_config/mcuboot_config.h"
